Move AlertsModal to separate file
