+++ components/Market/MarketNav/AlertsModal/AlertsModal.module.scss
